prep: Also extract buildId helper as metadata's plain function
The platform check will crash on metadata module import, while it
used to crash on Prep instanciation, but we don't expect third-party
windows code to use cozy-desktop as a lib anyway.
