+++ test/unit/prep.js
@@ -15,30 +15,6 @@ describe('Prep', function () {
   })
 
   describe('Helpers', function () {
-    describe('buildId', function () {
-      it('is available', function () {
-        let doc = {path: 'FOO'}
-        this.prep.buildId(doc)
-        doc._id.should.equal('FOO')
-      })
-
-      if (['linux', 'freebsd', 'sunos'].includes(process.platform)) {
-        it('is case insensitive on UNIX', function () {
-          let doc = {path: 'foo/bar/café'}
-          this.prep.buildId(doc)
-          doc._id.should.equal('foo/bar/café')
-        })
-      }
-
-      if (process.platform === 'darwin') {
-        it('is case sensitive on OSX', function () {
-          let doc = {path: 'foo/bar/café'}
-          this.prep.buildId(doc)
-          doc._id.should.equal('FOO/BAR/CAFÉ')
-        })
-      }
-    })
-
     describe('moveDoc', function () {
       it('calls moveFile for a file', function (done) {
         let doc = {
